+++ electron/preload.cjs
@@ -62,5 +62,6 @@ contextBridge.exposeInMainWorld("openarc", {
     list: (folderId) => ipcRenderer.invoke("files:list", folderId),
     rename: (folderId, id, name) => ipcRenderer.invoke("files:rename", { folderId, id, name }),
     remove: (folderId, id) => ipcRenderer.invoke("files:remove", { folderId, id }),
+    read: (folderId, id) => ipcRenderer.invoke("files:read", { folderId, id }),
   },
 });